Feature/HBAI-339: Support native func calling (#72)
* HBAI-339 Fix multi-byte utf decoding.
Simplify text gen logic.

* HBAI-339 Update repo id for llama.cpp

* HBAI-339 Change strings for {prompt}. Remove other dupe constants.
Fix eos token stopping.
Read "tags" to to install config.
Fix text_completion() not awaited.
Support the use of native tool defs.

* HBAI-339 Edit res msg

* HBAI-339 Put request_queue at api level.
Pass request obj to llama calls for interrupt.
Add /stop endpoint.
Fix stopping/pausing text generation.

* HBAI-339 Update prompt formats for Functionary.
Add "toolSchemaType" to Functionary model config.

* HBAI-339 Implement native tool calling

* HBAI-339 Add support for native function calling

* HBAI-339 Fix double streamed response.
Fix streaming events back to client.

* HBAI-339 Fix empty tool response data.
Terminate cli after each /completion response.

* HBAI-339 Up ver

* HBAI-339 Remove comment

* HBAI-339 Reset process var

* HBAI-339 Remove class members

+++ backends/api_server.py
@@ -3,6 +3,7 @@ import signal
 import sys
 import json
 import uvicorn
+import asyncio
 import httpx
 from collections.abc import Callable
 from fastapi import (
@@ -86,6 +87,7 @@ class ApiServer:
             print(f"{common.PRNT_API} Lifespan startup", flush=True)
             # Initialize global data here
             app.state.api = self
+            app.state.request_queue = asyncio.Queue()
             app.state.db_client = None
             app.state.llm = None  # Set each time user loads a model
             app.state.embed_model = None
@@ -121,8 +123,9 @@ class ApiServer:
     def shutdown(self, *args):
         try:
             print(f"{common.PRNT_API} Server forced to shutdown.", flush=True)
-            self.app.state.llm.unload()
-            os.kill(os.getpid(), signal.SIGTERM)  # or SIGINT
+            if self.app.state.llm:
+                self.app.state.llm.unload()
+                os.kill(os.getpid(), signal.SIGTERM)  # or SIGINT
         except Exception as e:
             print(
                 f"{common.PRNT_API} Failed to shutdown API server. Error: {e}",
